The Q format is an unsigned long long, not a uint64_t.

+++ examples/c/ex_schema.c
@@ -42,7 +42,7 @@ const char *home = "WT_TEST";
 typedef struct {
 	char country[5];
 	uint16_t year;
-	uint64_t population;
+	unsigned long long population;
 } POP_RECORD;
 /*! [schema decl] */
